move index creation into put job action
