+++ x-pack/plugin/ml-feature-index-builder/src/main/java/org/elasticsearch/xpack/ml/featureindexbuilder/action/TransportPutFeatureIndexBuilderJobAction.java
@@ -7,9 +7,11 @@
 package org.elasticsearch.xpack.ml.featureindexbuilder.action;
 
 import org.elasticsearch.action.ActionListener;
+import org.elasticsearch.action.admin.indices.create.CreateIndexRequest;
 import org.elasticsearch.action.support.ActionFilters;
 import org.elasticsearch.action.support.master.TransportMasterNodeAction;
 import org.elasticsearch.client.Client;
+import org.elasticsearch.client.IndicesAdminClient;
 import org.elasticsearch.cluster.ClusterState;
 import org.elasticsearch.cluster.block.ClusterBlockException;
 import org.elasticsearch.cluster.block.ClusterBlockLevel;
@@ -17,6 +19,7 @@ import org.elasticsearch.cluster.metadata.IndexNameExpressionResolver;
 import org.elasticsearch.cluster.service.ClusterService;
 import org.elasticsearch.common.inject.Inject;
 import org.elasticsearch.common.settings.Settings;
+import org.elasticsearch.common.xcontent.XContentType;
 import org.elasticsearch.license.LicenseUtils;
 import org.elasticsearch.license.XPackLicenseState;
 import org.elasticsearch.persistent.PersistentTasksService;
@@ -29,8 +32,14 @@ import org.elasticsearch.xpack.ml.featureindexbuilder.action.PutFeatureIndexBuil
 import org.elasticsearch.xpack.ml.featureindexbuilder.job.FeatureIndexBuilderJob;
 import org.elasticsearch.xpack.ml.featureindexbuilder.job.FeatureIndexBuilderJobConfig;
 
+import static org.elasticsearch.xpack.core.ml.job.persistence.ElasticsearchMappings.DOC_TYPE;
+
 public class TransportPutFeatureIndexBuilderJobAction
         extends TransportMasterNodeAction<PutFeatureIndexBuilderJobAction.Request, PutFeatureIndexBuilderJobAction.Response> {
+
+    // TODO: hack, to be replaced
+    private static final String PIVOT_INDEX = "pivot-reviews";
+
     private final XPackLicenseState licenseState;
     private final PersistentTasksService persistentTasksService;
     private final Client client;
@@ -67,7 +76,7 @@ public class TransportPutFeatureIndexBuilderJobAction
         XPackPlugin.checkReadyForXPackCustomMetadata(clusterState);
 
         FeatureIndexBuilderJob job = createFeatureIndexBuilderJob(request.getConfig(), threadPool);
-
+        createIndex(client, job.getConfig().getId());
         startPersistentTask(job, listener, persistentTasksService);
     }
 
@@ -89,5 +98,35 @@ public class TransportPutFeatureIndexBuilderJobAction
     @Override
     protected ClusterBlockException checkBlock(PutFeatureIndexBuilderJobAction.Request request, ClusterState state) {
         return state.blocks().globalBlockedException(ClusterBlockLevel.METADATA_WRITE);
+    }
+
+    /*
+     * Mocked demo case
+     *
+     * TODO: everything below will be replaced with proper implementation read from job configuration 
+     */
+    private static void createIndex(Client client, String suffix) {
+
+        String indexName = PIVOT_INDEX + "_" + suffix;
+        CreateIndexRequest request = new CreateIndexRequest(indexName);
+
+        request.settings(Settings.builder() // <1>
+                .put("index.number_of_shards", 1).put("index.number_of_replicas", 0));
+        request.mapping(DOC_TYPE, // <1>
+                "{\n" +
+                "  \"" + DOC_TYPE + "\": {\n" +
+                "    \"properties\": {\n" +
+                "      \"reviewerId\": {\n" +
+                "        \"type\": \"keyword\"\n" +
+                "      },\n" +
+                "      \"avg_rating\": {\n" +
+                "        \"type\": \"integer\"\n" +
+                "      }\n" +
+                "    }\n" +
+                "  }\n" +
+                "}", // <2>
+                XContentType.JSON);
+        IndicesAdminClient adminClient = client.admin().indices();
+        adminClient.create(request).actionGet();
     }
 }
